% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_in_one.R
\name{four_in_one}
\alias{four_in_one}
\title{Create a Four-in-One Residuals Plot}
\usage{
four_in_one(model, plotcolor, npplinecolor, yintcolor)
}
\description{
This function creates a set of residual plots which can
be used to ascertain the applicability of a given model
to a data set. The function's output and the included plots
are based off of Minitab®'s "Four in One" residual plots and
include the following:
}
\details{
\itemize{
\item A normal quantile plot
\item A residual plot comparing predicted residuals to their
actual value
\item A residual plot comparing predicted residuals to the order
in whihc they are observed
\item A humble histogram
}
}
