\name{four_in_one}
\alias{four_in_one}

\title{Create a four in one plot!

}
\description{
Provide a y variable and a x variable, preferably from a timeseries.
}
\usage{

}

\arguments{

}
\details{

}
\value{


}
\references{
}
\author{

}
\note{

}

\seealso{

}
\examples{
library(fourinone)
four_in_one(y=economics$pce,x=economics$psavert)
}
